+++ tests/pipeline/pipeline_input/response.json
