+++ tests/pipeline/pipeline_instance/response.json
