+++ tests/test_utils.py
@@ -1,0 +1,123 @@
+from tests.auth.client_credentials_test_utils import from_cache_not_expired_token
+
+well_known_response = {
+    "issuer": "https://localhost/auth/realms/api-catalog",
+    "authorization_endpoint": "https://localhost/auth/realms/api-catalog/protocol/openid-connect/auth",
+    "token_endpoint": "https://localhost/auth/realms/api-catalog/protocol/openid-connect/token",
+    "token_introspection_endpoint": "https://localhost/auth/realms/api-catalog/protocol/openid-connect/token/introspect",
+    "userinfo_endpoint": "https://localhost/auth/realms/api-catalog/protocol/openid-connect/userinfo",
+    "end_session_endpoint": "https://localhost/auth/realms/api-catalog/protocol/openid-connect/logout",
+    "jwks_uri": "https://localhost/auth/realms/api-catalog/protocol/openid-connect/certs",
+    "check_session_iframe": "https://localhost/auth/realms/api-catalog/protocol/openid-connect/login-status-iframe.html",
+    "grant_types_supported": [
+        "authorization_code",
+        "implicit",
+        "refresh_token",
+        "password",
+        "client_credentials",
+    ],
+    "response_types_supported": [
+        "code",
+        "none",
+        "id_token",
+        "token",
+        "id_token token",
+        "code id_token",
+        "code token",
+        "code id_token token",
+    ],
+    "subject_types_supported": ["public", "pairwise"],
+    "id_token_signing_alg_values_supported": [
+        "PS384",
+        "RS384",
+        "ES384",
+        "HS256",
+        "HS512",
+        "ES256",
+        "RS256",
+        "HS384",
+        "ES512",
+        "PS256",
+        "PS512",
+        "RS512",
+    ],
+    "id_token_encryption_alg_values_supported": ["RSA-OAEP", "RSA1_5"],
+    "id_token_encryption_enc_values_supported": ["A128CBC-HS256", "A128GCM"],
+    "userinfo_signing_alg_values_supported": [
+        "PS384",
+        "RS384",
+        "ES384",
+        "HS256",
+        "HS512",
+        "ES256",
+        "RS256",
+        "HS384",
+        "ES512",
+        "PS256",
+        "PS512",
+        "RS512",
+        "none",
+    ],
+    "request_object_signing_alg_values_supported": [
+        "PS384",
+        "ES384",
+        "RS384",
+        "ES256",
+        "RS256",
+        "ES512",
+        "PS256",
+        "PS512",
+        "RS512",
+        "none",
+    ],
+    "response_modes_supported": ["query", "fragment", "form_post"],
+    "registration_endpoint": "https://localhost/auth/realms/api-catalog/clients-registrations/openid-connect",
+    "token_endpoint_auth_methods_supported": [
+        "private_key_jwt",
+        "client_secret_basic",
+        "client_secret_post",
+        "client_secret_jwt",
+    ],
+    "token_endpoint_auth_signing_alg_values_supported": ["RS256"],
+    "claims_supported": [
+        "aud",
+        "sub",
+        "iss",
+        "auth_time",
+        "name",
+        "given_name",
+        "family_name",
+        "preferred_username",
+        "email",
+    ],
+    "claim_types_supported": ["normal"],
+    "claims_parameter_supported": False,
+    "scopes_supported": [
+        "openid",
+        "profile",
+        "email",
+        "address",
+        "phone",
+        "offline_access",
+        "roles",
+        "web-origins",
+        "microprofile-jwt",
+    ],
+    "request_parameter_supported": True,
+    "request_uri_parameter_supported": True,
+    "code_challenge_methods_supported": ["plain", "S256"],
+    "tls_client_certificate_bound_access_tokens": True,
+    "introspection_endpoint": "https://localhost/auth/realms/api-catalog/protocol/openid-connect/token/introspect",
+}
+
+
+client_credentials_response = {
+    "access_token": from_cache_not_expired_token,
+    "expires_in": 300,
+    "refresh_expires_in": 1800,
+    "refresh_token": from_cache_not_expired_token,
+    "token_type": "bearer",
+    "not-before-policy": 1563194597,
+    "session_state": "28957acd-9d5d-44f1-9f3f-7900b63fb010",
+    "scope": "profile email",
+}
